% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{associate_website_certificate_authority}
\alias{associate_website_certificate_authority}
\title{Imports the root certificate of a certificate authority (CA) used to obtain TLS certificates used by associated websites within the company network}
\usage{
associate_website_certificate_authority(FleetArn, Certificate,
  DisplayName = NULL)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{Certificate}{[required] The root certificate of the CA.}

\item{DisplayName}{The certificate name to display.}
}
\description{
Imports the root certificate of a certificate authority (CA) used to obtain TLS certificates used by associated websites within the company network.
}
\section{Accepted Parameters}{
\preformatted{associate_website_certificate_authority(
  FleetArn = "string",
  Certificate = "string",
  DisplayName = "string"
)
}
}

