% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{create_fleet}
\alias{create_fleet}
\title{Creates a fleet}
\usage{
create_fleet(FleetName, DisplayName = NULL,
  OptimizeForEndUserLocation = NULL)
}
\arguments{
\item{FleetName}{[required] A unique name for the fleet.}

\item{DisplayName}{The fleet name to display.}

\item{OptimizeForEndUserLocation}{The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.}
}
\description{
Creates a fleet. A fleet consists of resources and the configuration that delivers associated websites to authorized users who download and set up the Amazon WorkLink app.
}
\section{Accepted Parameters}{
\preformatted{create_fleet(
  FleetName = "string",
  DisplayName = "string",
  OptimizeForEndUserLocation = TRUE|FALSE
)
}
}

