% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{delete_fleet}
\alias{delete_fleet}
\title{Deletes a fleet}
\usage{
delete_fleet(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Deletes a fleet. Prevents users from accessing previously associated websites.
}
\section{Accepted Parameters}{
\preformatted{delete_fleet(
  FleetArn = "string"
)
}
}

