% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_audit_stream_configuration}
\alias{describe_audit_stream_configuration}
\title{Describes the configuration for delivering audit streams to the customer account}
\usage{
describe_audit_stream_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the configuration for delivering audit streams to the customer account.
}
\section{Accepted Parameters}{
\preformatted{describe_audit_stream_configuration(
  FleetArn = "string"
)
}
}

