% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_company_network_configuration}
\alias{describe_company_network_configuration}
\title{Describes the networking configuration to access the internal websites associated with the specified fleet}
\usage{
describe_company_network_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the networking configuration to access the internal websites associated with the specified fleet.
}
\section{Accepted Parameters}{
\preformatted{describe_company_network_configuration(
  FleetArn = "string"
)
}
}

