% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_device}
\alias{describe_device}
\title{Provides information about a user's device}
\usage{
describe_device(FleetArn, DeviceId)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{DeviceId}{[required] A unique identifier for a registered user's device.}
}
\description{
Provides information about a user's device.
}
\section{Accepted Parameters}{
\preformatted{describe_device(
  FleetArn = "string",
  DeviceId = "string"
)
}
}

