% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_device_policy_configuration}
\alias{describe_device_policy_configuration}
\title{Describes the device policy configuration for the specified fleet}
\usage{
describe_device_policy_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the device policy configuration for the specified fleet.
}
\section{Accepted Parameters}{
\preformatted{describe_device_policy_configuration(
  FleetArn = "string"
)
}
}

