% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_fleet_metadata}
\alias{describe_fleet_metadata}
\title{Provides basic information for the specified fleet, excluding identity provider, networking, and device configuration details}
\usage{
describe_fleet_metadata(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Provides basic information for the specified fleet, excluding identity provider, networking, and device configuration details.
}
\section{Accepted Parameters}{
\preformatted{describe_fleet_metadata(
  FleetArn = "string"
)
}
}

