% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_identity_provider_configuration}
\alias{describe_identity_provider_configuration}
\title{Describes the identity provider configuration of the specified fleet}
\usage{
describe_identity_provider_configuration(FleetArn)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}
}
\description{
Describes the identity provider configuration of the specified fleet.
}
\section{Accepted Parameters}{
\preformatted{describe_identity_provider_configuration(
  FleetArn = "string"
)
}
}

