% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{describe_website_certificate_authority}
\alias{describe_website_certificate_authority}
\title{Provides information about the certificate authority}
\usage{
describe_website_certificate_authority(FleetArn, WebsiteCaId)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{WebsiteCaId}{[required] A unique identifier for the certificate authority.}
}
\description{
Provides information about the certificate authority.
}
\section{Accepted Parameters}{
\preformatted{describe_website_certificate_authority(
  FleetArn = "string",
  WebsiteCaId = "string"
)
}
}

