% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{disassociate_website_certificate_authority}
\alias{disassociate_website_certificate_authority}
\title{Removes a certificate authority (CA)}
\usage{
disassociate_website_certificate_authority(FleetArn, WebsiteCaId)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{WebsiteCaId}{[required] A unique identifier for the CA.}
}
\description{
Removes a certificate authority (CA).
}
\section{Accepted Parameters}{
\preformatted{disassociate_website_certificate_authority(
  FleetArn = "string",
  WebsiteCaId = "string"
)
}
}

