% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{list_devices}
\alias{list_devices}
\title{Retrieves a list of devices registered with the specified fleet}
\usage{
list_devices(FleetArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{NextToken}{The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}

\item{MaxResults}{The maximum number of results to be included in the next page.}
}
\description{
Retrieves a list of devices registered with the specified fleet.
}
\section{Accepted Parameters}{
\preformatted{list_devices(
  FleetArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

