% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{list_fleets}
\alias{list_fleets}
\title{Retrieves a list of fleets for the current account and Region}
\usage{
list_fleets(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.}

\item{MaxResults}{The maximum number of results to be included in the next page.}
}
\description{
Retrieves a list of fleets for the current account and Region.
}
\section{Accepted Parameters}{
\preformatted{list_fleets(
  NextToken = "string",
  MaxResults = 123
)
}
}

