% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_package.R
\docType{package}
\name{paws.worklink-package}
\alias{paws.worklink}
\alias{paws.worklink-package}
\title{paws.worklink: Amazon WorkLink}
\description{
Amazon WorkLink is a cloud-based service that
    provides secure access to internal websites and web apps from iOS
    phones. In a single step, your users, such as employees, can access
    internal websites as efficiently as they access any other public
    website. They enter a URL in their web browser, or choose a link to an
    internal website in an email. Amazon WorkLink authenticates the
    user\'s access and securely renders authorized internal web content in
    a secure rendering service in the AWS cloud. Amazon WorkLink doesn\'t
    download or store any internal web content on mobile devices.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
