% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{sign_out_user}
\alias{sign_out_user}
\title{Signs the user out from all of their devices}
\usage{
sign_out_user(FleetArn, Username)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{Username}{[required] The name of the user.}
}
\description{
Signs the user out from all of their devices. The user can sign in again if they have valid credentials.
}
\section{Accepted Parameters}{
\preformatted{sign_out_user(
  FleetArn = "string",
  Username = "string"
)
}
}

