% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{update_audit_stream_configuration}
\alias{update_audit_stream_configuration}
\title{Updates the audit stream configuration for the fleet}
\usage{
update_audit_stream_configuration(FleetArn, AuditStreamArn = NULL)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{AuditStreamArn}{The ARN of the Amazon Kinesis data stream that receives the audit events.}
}
\description{
Updates the audit stream configuration for the fleet.
}
\section{Accepted Parameters}{
\preformatted{update_audit_stream_configuration(
  FleetArn = "string",
  AuditStreamArn = "string"
)
}
}

