% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{update_company_network_configuration}
\alias{update_company_network_configuration}
\title{Updates the company network configuration for the fleet}
\usage{
update_company_network_configuration(FleetArn, VpcId, SubnetIds,
  SecurityGroupIds)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{VpcId}{[required] The VPC with connectivity to associated websites.}

\item{SubnetIds}{[required] The subnets used for X-ENI connections from Amazon WorkLink rendering containers.}

\item{SecurityGroupIds}{[required] The security groups associated with access to the provided subnets.}
}
\description{
Updates the company network configuration for the fleet.
}
\section{Accepted Parameters}{
\preformatted{update_company_network_configuration(
  FleetArn = "string",
  VpcId = "string",
  SubnetIds = list(
    "string"
  ),
  SecurityGroupIds = list(
    "string"
  )
)
}
}

