% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{update_device_policy_configuration}
\alias{update_device_policy_configuration}
\title{Updates the device policy configuration for the fleet}
\usage{
update_device_policy_configuration(FleetArn, DeviceCaCertificate = NULL)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{DeviceCaCertificate}{The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.}
}
\description{
Updates the device policy configuration for the fleet.
}
\section{Accepted Parameters}{
\preformatted{update_device_policy_configuration(
  FleetArn = "string",
  DeviceCaCertificate = "string"
)
}
}

