% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{update_fleet_metadata}
\alias{update_fleet_metadata}
\title{Updates fleet metadata, such as DisplayName}
\usage{
update_fleet_metadata(FleetArn, DisplayName = NULL,
  OptimizeForEndUserLocation = NULL)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{DisplayName}{The fleet name to display. The existing DisplayName is unset if null is passed.}

\item{OptimizeForEndUserLocation}{The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.}
}
\description{
Updates fleet metadata, such as DisplayName.
}
\section{Accepted Parameters}{
\preformatted{update_fleet_metadata(
  FleetArn = "string",
  DisplayName = "string",
  OptimizeForEndUserLocation = TRUE|FALSE
)
}
}

