% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.worklink_operations.R
\name{update_identity_provider_configuration}
\alias{update_identity_provider_configuration}
\title{Updates the identity provider configuration for the fleet}
\usage{
update_identity_provider_configuration(FleetArn, IdentityProviderType,
  IdentityProviderSamlMetadata = NULL)
}
\arguments{
\item{FleetArn}{[required] The ARN of the fleet.}

\item{IdentityProviderType}{[required] The type of identity provider.}

\item{IdentityProviderSamlMetadata}{The SAML metadata document provided by the customer's identity provider. The existing IdentityProviderSamlMetadata is unset if null is passed.}
}
\description{
Updates the identity provider configuration for the fleet.
}
\section{Accepted Parameters}{
\preformatted{update_identity_provider_configuration(
  FleetArn = "string",
  IdentityProviderType = "SAML",
  IdentityProviderSamlMetadata = "string"
)
}
}

